requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Email
aiosmtplib==3.0.1

# SMS providers
httpx==0.26.0
twilio==8.11.0

# PDF Generation
reportlab==4.0.9
Pillow>=10.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
